+++ HISTORY.rst
@@ -10,11 +10,11 @@ TODO
 * skip already processed file from import
 * Breed name should be indexed as text
 * Breed name should be a unique key within species
-* Deal with ``mother_id`` and ``father_id`` (search for ``smarter_id`` in database)
 
 0.2.2.dev0
 ----------
 
+* Deal with ``mother_id`` and ``father_id`` (search for ``smarter_id`` in database)
 * Deal with multi-countries dataset
   - track country in aliases while importing breeds from dataset